factor out profile photo picker into new View
